backend/requirements.txt
flask
flask-cors
langchain-google-genai
langchain-core
langchain-community
faiss-cpu
whisper
python-dotenv
